unit, service: allow rerunning reload tasks.
In case a "reload" job cannot be cleanly merged to a previous job of the
same type, the job state is set to be "waiting". This triggers
reprocessing which doesn't lead to anything due to safeguards in unit.c
and service.c. Relax the conditions so that reload jobs can be rerun
when merge isn't possible.

Upstream-status: Submitted [https://github.com/systemd/systemd/pull/5839]

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>

%% original patch: 0001-unit-service-allow-rerunning-reload-tasks.patch