resolved: exclude DNS-SD PTR RRs from conflict resolution
According to RFC6763 p4.1 PTR resource records are used for
Service Instance Enumeration. In a local network multiple hosts
can provide services of the same type and it's up to a user
to choose which service instance to use. Therefore such
resource records of the same type (e.g. PTR resource records
for _ssh._tcp.local) are not in conflict. The same applies to
special meta records for _services._dns-sd._udp.local.

Never mark such resource records as conflicting.